refactor : Moved the thread related services in the threads package.
