+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/controllers/LikeController.java
@@ -2,7 +2,7 @@ package dev.anirban.harmoniq_backend.controllers;
 
 import dev.anirban.harmoniq_backend.constants.UrlConstants;
 import dev.anirban.harmoniq_backend.dto.common.ResponseWrapper;
-import dev.anirban.harmoniq_backend.service.LikeService;
+import dev.anirban.harmoniq_backend.service.threads.LikeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
