+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/service/threads/TagService.java
@@ -1,4 +1,4 @@
-package dev.anirban.harmoniq_backend.service;
+package dev.anirban.harmoniq_backend.service.threads;
 
 import dev.anirban.harmoniq_backend.entity.Tag;
 import dev.anirban.harmoniq_backend.repo.TagRepository;
